Split the mmu2 config into MMU2 and MMU2S
