+++ Firmware/mmu2_config.h
@@ -1,0 +1,6 @@
+/// @file
+#pragma once
+
+// Include the correct MMU2 config based on the printer variant.
+#include "Configuration_var.h"
+#include MMU_CONFIG_FILE
