+++ Firmware/variants/1_75mm_MK25S-RAMBo13a-E3Dv6full.h
@@ -503,6 +503,8 @@
 
 //#define SUPPORT_VERBOSITY
 
+#define MMU_CONFIG_FILE "mmu2/variants/config_MMU2S.h"
+
 #define MMU_FILAMENT_COUNT 5
 
 #define MMU_REQUIRED_FW_BUILDNR 132
