+++ Firmware/variants/1_75mm_MK3S-EINSy10a-E3Dv6full.h
@@ -680,6 +680,8 @@
 
 //#define SUPPORT_VERBOSITY
 
+#define MMU_CONFIG_FILE "mmu2/variants/config_MMU2S.h"
+
 #define MMU_FILAMENT_COUNT 5
 
 #define MMU_REQUIRED_FW_BUILDNR 83
